pyproject.toml
[project]
name = "iot-admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
]
